finish up feedback
