+++ src/pages/SigningOfficialDataSubmitters.js
@@ -5,7 +5,7 @@ import {Styles} from '../libs/theme';
 import {User} from '../libs/ajax';
 import { USER_ROLES } from '../libs/utils';
 import DataCustodianTable from '../components/data_custodian_table/DataCustodianTable';
-import SigningOfficialDAAPopup from '../components/SigningOfficialDaaAgreement';
+import SigningOfficialDaaAgreementWrapper from '../components/SigningOfficialDaaAgreementWrapper';
 
 
 export default function SigningOfficialConsole() {
@@ -44,9 +44,10 @@ export default function SigningOfficialConsole() {
   return (
     div({style: Styles.PAGE}, [
       div({style: {}, className: 'signing-official-tabs'}, [
-        //NOTE: Links to this custodian table have been removed, we are retaining it with the intention of repurposing it for data submitters
-        h(DataCustodianTable, {researchers, signingOfficial, unregisteredResearchers, isLoading}, []),
-        h(SigningOfficialDAAPopup, {}),
+        h(SigningOfficialDaaAgreementWrapper, {}, [
+          //NOTE: Links to this custodian table have been removed, we are retaining it with the intention of repurposing it for data submitters
+          h(DataCustodianTable, {researchers, signingOfficial, unregisteredResearchers, isLoading}, []),
+        ]),
       ])
     ])
   );
