+++ src/pages/SigningOfficialResearchers.js
@@ -5,7 +5,7 @@ import {Styles} from '../libs/theme';
 import SigningOfficialTable from '../components/signing_official_table/SigningOfficialTable';
 import {User} from '../libs/ajax';
 import { USER_ROLES } from '../libs/utils';
-import SigningOfficialDAAPopup from '../components/SigningOfficialDaaAgreement';
+import SigningOfficialDaaAgreementWrapper from '../components/SigningOfficialDaaAgreementWrapper';
 
 
 export default function SigningOfficialResearchers() {
@@ -37,9 +37,9 @@ export default function SigningOfficialResearchers() {
   return (
     div({style: Styles.PAGE}, [
       div({style: {}, className: 'signing-official-tabs'}, [
-        h(SigningOfficialTable, {researchers, signingOfficial, isLoading}, []),
-        // will automatically pop up if SO has not signed
-        h(SigningOfficialDAAPopup, {})
+        h(SigningOfficialDaaAgreementWrapper, {}, [
+          h(SigningOfficialTable, {researchers, signingOfficial, isLoading}, []),
+        ])
       ])
     ])
   );
